+++ CHANGELOG.md
@@ -21,6 +21,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
     - `RollingMacroFBeta`
     - `RollingMicroFBeta`
     - `RollingMultiFBeta`
+- `Multinomial` to `proba`
 
 ### Modified